removed jumbotron, refactored sql queries into functions, added login btn to navbar
