--- blank_page.php
@@ -1,11 +1,0 @@
-<?php 
-
-    $page_title = 'Word Find';
-    $subtitle = 'Search';
-    include 'inc/header.php';
-
-?>
-
-</div>
-
-<?php include('inc/footer.php'); ?>
