+++ db.php
@@ -10,6 +10,4 @@
 
     // Create PDO instance
     $pdo = new PDO($dsn, $user, $pass);
-    $pdo->setAttribute(PDO::ATTR_DEFAULT_FETCH_MODE, PDO::FETCH_OBJ);
-
-    
+    $pdo->setAttribute(PDO::ATTR_DEFAULT_FETCH_MODE, PDO::FETCH_OBJ);
