+++ includes/header.php
@@ -1,5 +1,15 @@
-<?php require_once 'db.php'; ?>
-<?php require_once 'indic-wp/word_processor.php'; ?>
+<?php
+
+    session_start();
+    require_once 'db.php';
+    require_once 'helpers/functions.php';
+    require_once 'indic-wp/word_processor.php';
+
+    if($page_title != 'Word Find Puzzle Maker'){
+        session_destroy();
+    }
+
+?>
 
 <!DOCTYPE html PUBLIC '-//W3C//DTD XHTML 1.0 Transitional//EN''http://www.w3.org/TR/xhtml1/DTD/xhtml1-transitional.dtd'><html xmlns='http://www.w3.org/1999/xhtml' xml:lang='en' lang='en'>
 <head>
@@ -22,12 +32,7 @@
 </head>
 
 <body class="d-flex flex-column">
-
-    <div class="container">
-        <a href="index.php"><header class="jumbotron jumbotron-fluid"></header></a>
-    </div>
-
-    <nav class="navbar navbar-expand-md navbar-dark bg-primary">
+    <nav class="navbar navbar-expand-lg navbar-dark bg-primary">
         <div class="container">
             <a class="navbar-brand" href="index.php"><img src="images/logo.png" alt="logo"></a>
             <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarNavAltMarkup" aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
@@ -38,8 +43,7 @@
                 <div class="navbar-nav">
                     <a class="nav-item nav-link<?php echo ($page_title == 'Word Find') ? ' active' : '' ; ?>" href="index.php">Home</a>
                     <a class="nav-item nav-link<?php echo ($page_title == 'Word Find Puzzle Maker') ? ' active' : '' ; ?>" href="create_puzzle.php">Create Puzzle</a>
-                    <!-- <a class="nav-item nav-link<?php // echo ($page_title == 'Admin') ? ' active' : '' ; ?>" href="admin.php">Admin</a> -->
-                    
+                    <a class="nav-item nav-link<?php echo ($page_title == 'login') ? ' active' : '' ; ?>" href="login.php">Login</a>
                 </div>
 
                 <form class="form-inline" action="search.php" method="post">
