+++ index.php
@@ -1,23 +1,14 @@
 <?php 
 
     $page_title = 'Word Find';
-    $subtitle = 'Popular Categories';
     include 'includes/header.php';
 
-    $sql = 'SELECT * FROM categories';
-    $stmt = $pdo->prepare($sql);
-    $stmt->execute();
-    $categories = $stmt->fetchAll();
+    $categories = getCategoriesWithTopNResults($pdo);
 
-
-    $sql2 = 'SELECT * FROM puzzles LEFT JOIN categories ON categories.cat_id = puzzles.cat_id';
-    $stmt2 = $pdo->prepare($sql2);
-    $stmt2->execute();
-    $puzzles = $stmt2->fetchAll();
 ?>
 
 <div class="card mt-4">
-    <h5 class="card-header"><?php echo $subtitle; ?></h5>
+    <h5 class="card-header">Popular Categories</h5>
 </div>
 
 <?php
@@ -25,11 +16,18 @@
     $numOfCols = 4;
     $rowCount = 0;
 
-    foreach ($categories as $cat):
+    foreach($categories as $category):
+
+        $data = [
+            'cat_name' => $category->cat_name,
+            'puzzle' => explode(',', $category->puzzle_names),
+            'puzzle_id' => explode(',', $category->puzzle_ids)
+        ];
+
         if($rowCount % $numOfCols == 0):
 ?>
         
-        <div class="row mt-3"> 
+        <div class="row"> 
         
     <?php
 
@@ -37,10 +35,10 @@
         $rowCount++;
 
     ?>  
-            <div class="col-md">
+            <div class="col-md mt-3">
                 <div class="card h-100">
-                    <h6 class="card-header"><a href="category.php?categoryId=<?php echo $cat->cat_id; ?>" class="card-link">
-                        <?php echo $cat->cat_name; ?>
+                    <h6 class="card-header"><a href="category.php?categoryId=<?php echo $category->cat_id; ?>" class="card-link">
+                        <?php echo $data['cat_name']; ?>
                         <svg width="1em" height="1em" viewBox="0 0 16 16" class="bi bi-three-dots" fill="currentColor" xmlns="http://www.w3.org/2000/svg">
                             <path fill-rule="evenodd" d="M3 9.5a1.5 1.5 0 1 1 0-3 1.5 1.5 0 0 1 0 3zm5 0a1.5 1.5 0 1 1 0-3 1.5 1.5 0 0 1 0 3zm5 0a1.5 1.5 0 1 1 0-3 1.5 1.5 0 0 1 0 3z"/>
                         </svg>
@@ -48,17 +46,10 @@
                     
                     <div class="card-body">
                         <?php
-                            foreach($puzzles as $puzzle): 
-                                if($cat->cat_id == $puzzle->cat_id):
-                        ?>
-                                    <p class="card-text"><a href="puzzle.php?puzzleId=<?php echo $puzzle->puzzle_id; ?>" class="card-link"><?php echo $puzzle->title; ?></a></p>
-
-                        <?php 
-
-                                endif; 
-                            endforeach;
-                            
-                        ?>
+                            foreach($data['puzzle'] as $index => $puzzle): ?>
+                                <p class="card-text"><a href="puzzle.php?puzzleId=<?php echo $data['puzzle_id'][$index]; ?>" class="card-link"><?php echo $puzzle; ?></a></p>
+                                
+                            <?php endforeach; ?>
                     </div>
                 </div>
             </div>
@@ -66,9 +57,12 @@
         <?php
             if($rowCount % $numOfCols == 0): ?>
         </div>
+
 <?php 
             endif;
-    endforeach; // end categories
+
+        endforeach;
+
 ?>
 
 </div>
