+++ template.php
@@ -1,0 +1,14 @@
+<?php 
+
+    $page_title = 'blank page template';
+    include 'includes/header.php';
+
+?>
+
+<div class="card mt-4">
+    <h5 class="card-header">Blank Page Template</h5>
+</div>
+    
+</div>
+
+<?php include('includes/footer.php'); ?>
